Optimization of classes and methods.

+++ java/src/bpa/BpaCostCalculator.java
@@ -4,7 +4,7 @@ import java.util.Map;
 
 public interface BpaCostCalculator {
 	
-	Integer getActivityCost(String activityName);
-	
+	//Integer getActivityCost(String activityName);
+	Double getActivityCost(String activityName);
 	
 }
